Feature yztemplate 20201016 zlt

See merge request ajy/yun_shop!633

+++ app/backend/modules/setting/controllers/DefaultNoticeController.php
@@ -38,17 +38,24 @@ class DefaultNoticeController extends BaseController
     public function index() {
         $notice_name = request()->notice_name;
         $setting_name = request()->setting_name;
+        $notice_id = request()->notice_id;
 
         $notice = \Setting::get($setting_name);//获取设置信息
 
-        $message_template = $this->MessageTempModel->getTempIdByNoticeType($notice_name);//获取消息通知模版
-        $has_template_id = $this->wechat_list->where('template_id',$message_template->template_id)->first();//查询是否存在template_id,不存在则新建
+        //fixby-zlt-templateset 2020-10-10
+        if(!empty($notice_id)){
+            $notice[$notice_name] = $notice_id;
+        }else{
+            $message_template = $this->MessageTempModel->getTempIdByNoticeType($notice_name);//获取消息通知模版
+            $has_template_id = $this->wechat_list->where('template_id',$message_template->template_id)->first();//查询是否存在template_id,不存在则新建
 
-        if ($has_template_id) {
-            $notice[$notice_name] = (string)$message_template->id;
-        } else {
-            $notice[$notice_name] = $this->createDefaultMessageTemp($notice_name);
+            if ($has_template_id) {
+                $notice[$notice_name] = (string)$message_template->id;
+            } else {
+                $notice[$notice_name] = $this->createDefaultMessageTemp($notice_name);
+            }
         }
+
         \Setting::set($setting_name, $notice);
         echo json_encode([
             'result' => '1',
@@ -70,18 +77,24 @@ class DefaultNoticeController extends BaseController
     public function store()
     {
         $notice_name = \YunShop::request()->notice_name;
+        $notice_id = request()->notice_id;
         $setting_name = \YunShop::request()->setting_name;
 
         $message_template = $this->MessageTempModel->getTempIdByNoticeType($notice_name);//获取消息通知模版
         $has_template_id = $this->wechat_list->where('template_id',$message_template->template_id)->first();//查询是否存在template_id,不存在则新建
 
-        if ($has_template_id){
-            $item = (string)$message_template->id;
-        } else {
-            if ($message_template) {
-                $message_template->delete();
+        //fixby-zlt-templateset 2020-10-13
+        if(!empty($notice_id)){
+            $item = $notice_id;
+        }else {
+            if ($has_template_id) {
+                $item = (string)$message_template->id;
+            } else {
+                if ($message_template) {
+                    $message_template->delete();
+                }
+                $item = $this->createDefaultMessageTemp($notice_name);
             }
-            $item = $this->createDefaultMessageTemp($notice_name);
         }
         \Setting::set($setting_name, $item);
         $setting = explode('.',$setting_name);
